fix: refactor method into a shared resource for reduced repetition
