+++ packages/itinerary-body/src/stories/OtpRrItineraryBody.story.tsx
@@ -1,7 +1,7 @@
-import { Itinerary, Leg } from "@opentripplanner/types";
+import { Itinerary } from "@opentripplanner/types";
 import React, { FunctionComponent, ReactElement } from "react";
+import RouteDescriptionFooterWithWaitTimes from "../ItineraryBody/components";
 
-import { differenceInMinutes } from "date-fns";
 import ItineraryBody from "..";
 import {
   CustomTimeColumnContent,
@@ -11,11 +11,10 @@ import OtpRRLineColumnContent from "../otp-react-redux/line-column-content";
 import { PlaceName as OtpRRPlaceName } from "../otp-react-redux";
 import OtpRRRouteDescription from "../otp-react-redux/route-description";
 import { isRunningJest } from "../../../../.storybook/react-intl";
-import { RouteDescriptionFooterProps, TimeColumnContentProps } from "../types";
+import { TimeColumnContentProps } from "../types";
 
 import ItineraryBodyDefaultsWrapper from "./itinerary-body-defaults-wrapper";
 import LegIconWithA11y from "./LegIconWithA11y";
-import { DefaultRouteDescriptionFooter } from "../defaults/route-description-footer";
 
 // import mock itinaries. These are all trip plan outputs from OTP.
 const bikeOnlyItinerary = require("../__mocks__/itineraries/bike-only.json");
@@ -53,17 +52,6 @@ interface StoryWrapperProps {
   TimeColumnContent: FunctionComponent<TimeColumnContentProps>;
   alwaysCollapseAlerts: boolean;
 }
-
-const RouteDescriptionFooterWithWaitTimes = ({ leg }: { leg: Leg }) => {
-  const toTime = leg.to.arrival || 0;
-  const fromTime = leg.from.arrival || 0;
-  const waitMinutes = differenceInMinutes(new Date(toTime), new Date(fromTime));
-  const TypedRouteDescriptionFooter = DefaultRouteDescriptionFooter as React.FC<
-    RouteDescriptionFooterProps
-  >;
-
-  return <TypedRouteDescriptionFooter leg={leg} waitMinutes={waitMinutes} />;
-};
 
 function OtpRRItineraryBodyWrapper({
   itinerary,
